+++ server/src/test/java/org/allaymc/api/utils/UtilsTest.java
@@ -1,7 +1,5 @@
 package org.allaymc.api.utils;
 
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.Unpooled;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -28,12 +26,5 @@ class UtilsTest {
     @Test
     void testAppendBytes() {
         assertEquals(new String(new byte[]{1, 2, 3, 4, 5, 6, 7, 8, 9}), new String(Utils.appendBytes(new byte[]{1, 2, 3}, new byte[]{4, 5, 6}, new byte[]{7, 8, 9})));
-    }
-
-    @Test
-    void testConvertByteBuf2Array() {
-        byte[] bytes = new byte[]{1, 2, 3, 4, 5, 6, 7, 8, 9};
-        ByteBuf buf = Unpooled.wrappedBuffer(bytes);
-        assertEquals(new String(bytes), new String(Utils.convertByteBuf2Array(buf)));
     }
 }